♻️ [refactor] 컴포넌트 분리
컴포넌트를 분리함

fixes: 추후 수정 가능성 있음
