+++ src/pages/Movie.jsx
@@ -1,8 +1,8 @@
 import { Helmet } from "react-helmet-async";
-import S from "./Movie.module.css";
 import { Link } from "react-router-dom";
 import React, { useRef, useEffect, useState } from "react";
-import { MovieNav, MovieContent } from "./MovieNav";
+import MovieNav from "./../components/category/MovieNav";
+import MovieContent from "./../components/category/MovieContent";
 
 export default function Movie() {
 	return (
