--- src/pages/Program.module.css
@@ -1,15 +1,0 @@
-@import "tailwindcss/base";
-@import "tailwindcss/components";
-@import "tailwindcss/utilities";
-
-.button {
-  @apply bg-transparent hover:bg-yellow-500 duration-300 text-white text-lg border border-yellow-200 rounded-full py-3 w-[80%];
-}
-
-.buttonActive {
-  @apply bg-yellow-500 text-white text-lg border border-yellow-200 rounded-full py-3 w-[80%];
-}
-
-.listContent {
-  @apply w-[20%] max-md:w-[30%] hover:-translate-y-2 duration-300 text-gray200 hover:text-white;
-}
